Bart/dp 6626 (#17)
* DP-6626 - Add import for oplaadpalen from Allego

From Allego we first get a complete list of all 'oplaadpalen' (in Amsterdam) with the status (Available or other)

The import form Allego can work in two ways. If the oplaadpalen table does not yet exist or
the --delete parameter is added, a new table is created. Otherwise the status for all existing
entries is updated, and new oplaadpalen are retrieved and create for at most --max_inserts entries

This was added in order to prevent to much calls to the Allego API

* DP-6626 - Add update for chargepoint details. Use environment variables in import shell script