Add geosearch output in GeoJSON format.
Currently this is wrapped in a pager that puts paged features in a result set.
This is different form the generic geosearch. Nevertheless this looks like a
improvement.

The generic geosearch returns GeoJSON without the geometry field. That is not
according to the GeoJSON standard. But we can mimic this behavious by adding
detail=0 parameter