Making evenementen import more generic
The 'generic' GeoJSON import and the evenementen import that makes use of ogr2ogr
are integrated, and import and checks are decoupled from the data definitions.

This process can also be used for other types of import (files, JSON, shapefiles)
but only the happy path for the data model of evenementen is implemented
currently.